fix(trusty-analyze): route Kotlin/Swift/Ruby/PHP files to their linters (closes #963)
`run_diagnostics` gates every file on `LanguageDetector::detect_file` before
dispatching to its linters, but `LANGUAGE_DETECTORS` only carried 7 detectors.
Files whose extension wasn't in that table (`.kt`/`.kts`/`.swift`/`.rb`/`.php`)
were `continue`-skipped before any linter ran, so detekt, swiftlint, rubocop,
and phpstan could never fire through the endpoint even when installed —
`diagnostics` returned `count: 0`, which reads as "clean code" rather than
"not run".

Add `detect_kotlin`/`detect_swift`/`detect_ruby`/`detect_php`, returning tags
(`kotlin`/`swift`/`ruby`/`php`) that match the `StaticTool::language()` bucket
keys, and append them to `LANGUAGE_DETECTORS`. (The C/C++ note in the issue is a
non-issue: clang-tidy already registers under `cpp`, matching `detect_cpp`.)

Detection remains purely extension-based — content/syntax is not inspected at
the routing gate; the tree-sitter adapters parse contents only afterward.

Because the feature pushed `detection.rs` over its frozen line-cap budget,
split it into `detection/mod.rs` (extension + build-system detection) and
`detection/frameworks.rs` (manifest-content framework inference), each well
under the 500-line cap, and ratchet the allowlist entry away.

Test: `detect_file_extension_mapping` now covers .kt/.kts/.swift/.rb/.php; new
`detect_file_returns_linter_bucket_keys_for_kotlin_swift_ruby_php` locks the
tags to the linter bucket keys.
